Cambios en el scroll.js
